+++ packages/extension-polkagate/src/components/FormatPrice.tsx
@@ -5,15 +5,17 @@ import type { BN } from '@polkadot/util';
 
 import { Grid, Skeleton, Typography } from '@mui/material';
 import React, { useMemo } from 'react';
+import CountUp from 'react-countup';
 
 import { useCurrency } from '../hooks';
 import { ASSETS_AS_CURRENCY_LIST } from '../util/currencyList';
-import { amountToHuman } from '../util/utils';
+import { amountToHuman, fixFloatingPoint } from '../util/utils';
 
 interface Props {
   amount?: BN | null;
   decimalPoint?: number;
   decimals?: number;
+  commify?: boolean;
   fontSize?: string;
   fontWeight?: number;
   lineHeight?: number;
@@ -26,6 +28,7 @@ interface Props {
   textColor?: string;
   height?: number;
   width?: string;
+  withCountUp?: boolean;
 }
 
 export function nFormatter (num: number, decimalPoint: number) {
@@ -53,7 +56,7 @@ export function nFormatter (num: number, decimalPoint: number) {
 
 const DECIMAL_POINTS_FOR_CRYPTO_AS_CURRENCY = 4;
 
-function FormatPrice ({ amount, decimalPoint = 2, decimals, fontSize, fontWeight, height, lineHeight = 1, mt = '0px', num, price, sign, skeletonHeight = 15, textAlign = 'left', textColor, width = '90px' }: Props): React.ReactElement<Props> {
+function FormatPrice ({ amount, commify, decimalPoint = 2, decimals, fontSize, fontWeight, height, lineHeight = 1, mt = '0px', num, price, sign, skeletonHeight = 15, textAlign = 'left', textColor, width = '90px', withCountUp }: Props): React.ReactElement<Props> {
   const currency = useCurrency();
 
   const total = useMemo(() => {
@@ -90,7 +93,17 @@ function FormatPrice ({ amount, decimalPoint = 2, decimals, fontSize, fontWeight
           lineHeight={lineHeight}
           sx={{ color: textColor }}
         >
-          {sign || currency?.sign || ''}{nFormatter(total as number, _decimalPoint)}
+          {withCountUp
+            ? <CountUp
+              decimals={_decimalPoint}
+              duration={1}
+              end={parseFloat(String(total))}
+              prefix={sign || currency?.sign || ''}
+            />
+            : <>
+              {sign || currency?.sign || ''}{ commify ? fixFloatingPoint(total as number, _decimalPoint, true) : nFormatter(total as number, _decimalPoint)}
+            </>
+          }
         </Typography>
         : <Skeleton
           animation='wave'