+++ packages/extension-polkagate/src/hooks/index.ts
@@ -117,6 +117,7 @@ export { default as useTransactionState } from './useTransactionState';
 export { default as useTranslation } from './useTranslation';
 export { default as useUnstakingAmount } from './useUnstakingAmount';
 export { default as useUnSupportedNetwork } from './useUnSupportedNetwork';
+export { default as useValidatorApy } from './useValidatorApy';
 export { default as useValidators } from './useValidators';
 export { default as useValidatorsIdentities } from './useValidatorsIdentities';
 export { default as useValidatorSuggestion } from './useValidatorSuggestion';