+++ packages/extension-polkagate/src/popup/home/ProfileTab.tsx
@@ -85,12 +85,14 @@ export default function ProfileTab ({ index, isContainerHovered, isSelected, ord
         orientation='horizontal'
         sx={{
           '&:hover': { boxShadow: shadowOnHover },
-          bgcolor: getProfileColor(index, theme) || 'background.paper',
+          border:1,
+          borderColor: getProfileColor(index, theme) || 'background.paper',
           borderRadius: '10px',
           boxShadow: shadow,
           cursor: 'pointer',
           height: COLLAPSED_SIZE,
-          my: '2px'
+          my: '2px',
+          opacity: isDarkMode ? (isContainerHovered || isSelected ? 1 : 0.3) : undefined
         }}
       >
         <Grid alignItems='center' container item justifyContent='center' sx={{ display: 'flex', flexDirection: 'row', flexWrap: 'nowrap', minWidth: '40px', px: '8px', width: 'fit-content' }}>