+++ packages/extension-polkagate/src/util/types.ts
@@ -73,6 +73,7 @@ export interface ValidatorInfo extends DeriveStakingQuery {
     others: Other[]
   };
   accountInfo?: DeriveAccountInfo;
+  apy?: string | null;
   isOversubscribed?: {
     notSafe: boolean;
     safe: boolean;